leet20 last considerations
As I thought, you're supposed to use a Stack.
Reading through the string, if it's an opening symbol, you append it to the stack. Otherwise you compare it with the last item in the stack [-1], if they are pairs, you remove it from the stack, else you return False. yo yo! If the stack is empty by the time you finsh going through the string, you can return True! yeah!!!! so cool!!!! I'm so tired, I wanna go to sleep.